[test] testContainer Redis 추가
