+++ .gitignore
@@ -27,7 +27,6 @@ out/
 !**/src/test/**/out/
 gradle/
 **/application-local.yml
-**/application-test.yml
 
 ### NetBeans ###
 /nbproject/private/
