+++ src/test/java/hhplus/tdd/concert/app/application/service/reservation/ReservationServiceIntegrationTest.java
@@ -214,57 +214,4 @@ public class ReservationServiceIntegrationTest {
         log.info("[낙관락_retry_멤버1000명이_1개좌석_쟁탈결과_1명만성공] 소요시간: {}s", secDiffTime);
     }
 
-    @Test
-    public void 레디스_PUB_SUB_멤버100명이_1개좌석_쟁탈결과_1명만성공() throws InterruptedException {
-        long beforeTime = System.currentTimeMillis();
-
-        // given
-        int totalTasks = 100;
-        int capacity = 1;
-        AtomicInteger failCnt = new AtomicInteger();
-        List<Member> memberList = new ArrayList<>();
-        List<Waiting> waitingList = new ArrayList<>();
-        IntStream.range(0, totalTasks).forEach(i -> {
-            Member member = Member.builder().memberName("김소리" + i).charge(15000).build();
-            memberList.add(member);
-            waitingList.add(Waiting.builder().token("sample-token" + i).member(member).status(WaitingStatus.ACTIVE).createAt(testBase.now.minusMinutes(30)).expiredAt(testBase.now.plusMinutes(30)).build());
-        });
-        memberRepository.saveAll(memberList);
-        waitingRepository.saveAll(waitingList);
-        concertSeatRepository.save(testBase.concertSeatStandBy);
-
-        CountDownLatch latch = new CountDownLatch(totalTasks);
-        ExecutorService executorService = Executors.newFixedThreadPool(totalTasks);
-
-        // when
-        for (Waiting waiting : waitingList) {
-            executorService.execute(() -> {
-                try {
-                    reserveDistributedLockAop.processReserveRedisPubSub(waiting.getToken(), testBase.concertSeatStandBy.getSeatId());
-                } catch (Exception e) {
-                    failCnt.getAndIncrement();
-                } finally {
-                    latch.countDown();
-                }
-            });
-        }
-
-        latch.await();
-
-        List<Reservation> reservation = reservationRepository.findAll();
-
-        // then
-        assertEquals(totalTasks - capacity, failCnt.get());
-        assertEquals(1, reservation.size());
-        assertEquals(ReserveStatus.PENDING, reservation.get(0).getReserveStatus());
-
-        executorService.shutdown();
-
-        long afterTime = System.currentTimeMillis();
-        double  secDiffTime = (afterTime - beforeTime)/1000.0;
-
-        log.info("[레디스_PUB_SUB_멤버100명이_1개좌석_쟁탈결과_1명만성공] 소요시간: {}s", secDiffTime);
-    }
-
-
 }
